ArrayIterator with ReadOnlyListIerator in new AbstractIterator
